requirements.txt
flask
google-generativeai
gTTS
fpdf
python-pptx
Pillow
SpeechRecognition
python-dotenv